feat: remove start_time from all workload types [DET-5979] (#2912)
This change removes start time for steps, validations and checkpoints and, as a result, obsoletes the hack code to derive start time, since it is no longer reported by the harness. Additionally, step ID was dropped and replaced by a plain old SERIAL since it is also no longer reported by the harness.

Originally after the harness stopped reporting start time, it was derived as a hack in a painful way (but accurately, for the time being) in an attempt to not break the reporting done by the historical resource allocation project, but this has been worked around permanently by aggregating over public.allocations instead. Step ID was never really needed, but was kept "just in case" until it was clear after push architecture it was not needed at all.