chore: unify task types [DET-5950, DET-5955] (#2938)
The change updates `command.go` and surrounding areas to use the new `task.Allocation` struct to manage its allocation. This has many benefits, mostly that, by default, all behaviors previously available to commands or trials separately are now available to both (only a config variable away). For example, this means trials would be able to enable a proxy or have some sort of idle timeout (maybe to detect hangs), commands can support preemption or rendezvous (for multi-agent commands), etc.

This is definitely not the "end all" for how this code will look. Probably over the course of fixing logging and adding persistence I'll move it around a lot more; but, I think it's generally in the right direction, and I'm trying to move fast in the right direction than slowly, perfectly.

What I mean by this is that within the job<->task<->allocation hierarchy, I've, for now, just pushed all functionality to the bottom (the "allocation"), though in the future it may make sense for some features to live in differently layers of the stack. Though, since there are no generic "jobs" or generic "tasks" yet to put shared task functionality, this is just easier for now. E.G. proxies really may make more sense at the task level, since it is the task that knows what kind of proxy it wants, but to reuse the code to register proxies I've pushed it down to allocations for now. In the future, it's possible that the "framework" will be that allocations bubble all allocation events up to tasks and tasks bubble all task events up to jobs, and at each layer some features sit and hook into these events, but it's just not quite all the way there yet.