refactor: pull out intermediary resources representation for RM (#3663)
This change introduces an intermediary representation for container,
pods or jobs called "Resources" which are an arbitrary-ish clumps of
resources from a resource manager.  This is the first change in a series
to allow us to integrate with slurm more naturally.  This reasons for
this change follow.  
1. When we converse with resource managers that do not speak terms of
containers the rename portion of this change is required else the code
reads like nonsense.
2. It provides an easier path to LCD shared resource manager features
that does not involve refactor the internals of Kubernetes or agents
resource managers.
3. It provides a path to put non-shared features behind clean feature
flags.
This approach takes our current interface and makes it flexible enough
to be generalized but as of now is mostly a noop, so to enable can work
with more resource managers without too much fuss.  There are a few
other approaches that were not chosen.  One would be to force each
resource manager we support to be compatible with our current interface
through a serious of shims/scripts on the resource manager side of the
interface.  This isn't really ideal, mostly because it isn't clear it's
even possible, much less sustainable with all resource managers we may
support.  Another would be to move the allocation actor into resource
managers and draw the line where the RM interface starts at a different
place.  This is promising, but then existing features have to be
refactor/re-implemented at a different layer and is generally a lot more
work.  Also the current approach is still useful with that approach and
may transpose to it later if it still seems appealing.