feat: break workload info from trial endpoint into a new endpoint [DET-6729] (#3635)
When a trial has run many workloads, the response to the trial details
endpoint for it can become very large and unwieldy. Since we don't
always need the full set of workloads, we add a new endpoint just for
providing workload information and add the most needed summary
information to the general trial endpoint. The original endpoint still
contains workloads for now, but that will change once the clients are
switched to the new endpoint.